Factor out proofs code into a separate package
